+++ include/dle/ghs_printer.h
@@ -33,13 +33,13 @@
  *   NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  * @file ghs_printer.h
- * @brief some conveneience implementations to print le::ghs::GhsState objects with std::ostream
+ * @brief some conveneience implementations to print dle::GhsState objects with std::ostream
  *
  */
 #ifndef ghs_printer_hpp
 #define ghs_printer_hpp
 
-#include "ghs/ghs.h"
+#include <dle/ghs.h>
 #include <string>
 #include <iostream>
 
@@ -55,6 +55,6 @@ std::string dump_edges(const le::ghs::GhsState<N,A>&) ;
 template <std::size_t N, std::size_t A>
 std::ostream& operator << ( std::ostream&, const le::ghs::GhsState<N,A>&);
 
-#include "ghs/ghs_printer_impl.hpp"
+#include <dle/ghs_printer_impl.hpp>
 
 #endif
