+++ include/dle/ghs_printer_impl.hpp
@@ -34,23 +34,17 @@
  * @brief the implementation of ghs_printer.h Just include ghs_printer.h
  *
  */
-#include "ghs/ghs_printer.h"
-
+#include <ghs/ghs_printer.h>
 #include <sstream>
-
-using le::ghs::Edge;
-using le::ghs::agent_t;
-using le::ghs::UNKNOWN;
-using le::ghs::MST;
-using le::ghs::MST_PARENT;
-using le::ghs::DELETED;
-using namespace le;
 
 /** 
  * Prints the edges in a neatly formatted way
  */
 template <std::size_t A, std::size_t B>
-std::string  dump_edges(const le::ghs::GhsState<A,B> &s) {
+std::string  dump_edges(const dle::GhsState<A,B> &s) {
+
+  using dle::Edge;
+
   std::stringstream ss;
   ss<<"( ";
   Edge mwoe = s.mwoe();
@@ -88,7 +82,7 @@ std::string  dump_edges(const le::ghs::GhsState<A,B> &s) {
  * Prints some basic GHS information, like leader, level, # waiting, # delayed, converged (yes/no)
  */
 template <std::size_t A, std::size_t B>
-std::ostream& operator << ( std::ostream& outs, const le::ghs::GhsState<A,B> & s){
+std::ostream& operator << ( std::ostream& outs, const dle::GhsState<A,B> & s){
   outs<<"{id:"<<s.get_id()<<" ";
   outs<<"leader:"<<s.get_leader_id()<<" ";
   outs<<"level:"<<s.get_level()<<" ";
